🔀 chore(route.ts): move authOptions to a separate file for better organization and reusability 🔀 chore(page.tsx): import authOptions from the new options file to use in getServerSession 🔀 chore(options.ts): create a new file to store authOptions for NextAuth configuration
