+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,22 +1,5 @@
+import { authOptions } from '@/resources/options'
 import NextAuth, { NextAuthOptions } from 'next-auth'
-import GoogleProvider from 'next-auth/providers/google'
-
-type ClientType = {
-  clientId: string
-  clientSecret: string
-}
-
-const authOptions: NextAuthOptions = {
-  debug: true,
-  session: { strategy: 'jwt' },
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET,
-    } as ClientType),
-  ],
-  secret: process.env.NEXTAUTH_SECRET,
-}
 
 const handler = NextAuth(authOptions)
