+++ src/app/page.tsx
@@ -1,11 +1,15 @@
 import { LoginButton, LogoutButton } from '@/components/molecules/LoginButton'
+import { authOptions } from '@/resources/options'
+import { getServerSession } from 'next-auth'
 
-export default function Home() {
+export default async function Home() {
+  const session = await getServerSession(authOptions)
+  const user = session?.user
+
   return (
     <div>
-      <p>hello world</p>
-      <LoginButton />
-      <LogoutButton />
+      <p>hello {user ? user?.name : 'world'}</p>
+      {user ? <LogoutButton /> : <LoginButton />}
     </div>
   )
 }
